+++ shesha-reactjs/src/designer-components/attachmentsEditor/attachmentsEditor.tsx
@@ -5,15 +5,14 @@ import React from 'react';
 import { CustomFile } from '@/components';
 import ConfigurableFormItem from '@/components/formDesigner/components/formItem';
 import { IToolboxComponent } from '@/interfaces';
-import { useForm, useFormData, useGlobalState, useHttpClient, useSheshaApplication } from '@/providers';
+import { useDataContextManagerActions, useForm, useFormData, useGlobalState, useHttpClient, useSheshaApplication } from '@/providers';
 import { IConfigurableFormComponent, IInputStyles } from '@/providers/form/models';
 import {
   evaluateValue,
-  executeScript,
+  executeScriptSync,
   validateConfigurableComponentSettings,
 } from '@/providers/form/utils';
 import StoredFilesProvider from '@/providers/storedFiles';
-import { IStoredFile } from '@/providers/storedFiles/contexts';
 import { getSettings } from './settings';
 import { migrateCustomFunctions, migratePropertyName, migrateReadOnly } from '@/designer-components/_common-migrations/migrateSettings';
 import { migrateVisibility } from '@/designer-components/_common-migrations/migrateVisibility';
@@ -36,6 +35,7 @@ export interface IAttachmentsEditorProps extends IConfigurableFormComponent, IIn
   isDragger?: boolean;
   maxHeight?: string;
   onFileChanged?: string;
+  onDownload?: string;
   downloadZip?: boolean;
   layout: layoutType;
   listType: listType;
@@ -43,6 +43,7 @@ export interface IAttachmentsEditorProps extends IConfigurableFormComponent, IIn
   thumbnailHeight?: string;
   borderRadius?: number;
   hideFileName?: boolean;
+  removeFieldFromPayload?: boolean;
 }
 
 const AttachmentsEditor: IToolboxComponent<IAttachmentsEditorProps> = {
@@ -57,33 +58,42 @@ const AttachmentsEditor: IToolboxComponent<IAttachmentsEditorProps> = {
     const { data } = useFormData();
     const { globalState, setState: setGlobalState } = useGlobalState();
     const { message } = App.useApp();
-
+    const pageContext = useDataContextManagerActions(false)?.getPageContext();
     const ownerId = evaluateValue(`${model.ownerId}`, { data: data, globalState });
 
     const enabled = !model.readOnly;
 
-    const onFileListChanged = (fileList: IStoredFile[]) => {
+    const executeScript = (script, value) => {
 
-      if (!model.onFileChanged)
-        return;
-
-      executeScript<void>(model.onFileChanged, {
-        fileList,
+      executeScriptSync(script, {
+        value: value,
         data,
         form: getFormApi(form),
         globalState,
         http: httpClient,
         message,
         moment,
-        setGlobalState
+        setGlobalState,
+        pageContext
       });
     };
-
+    
     return (
       // Add GHOST_PAYLOAD_KEY to remove field from the payload
       // File list uses propertyName only for support Required feature
-      <ConfigurableFormItem model={{ ...model, propertyName: `${GHOST_PAYLOAD_KEY}_${model.propertyName}` }}>
+      <ConfigurableFormItem model={{ ...model, propertyName: model.propertyName || `${GHOST_PAYLOAD_KEY}_${model.id}` }}>
         {(value, onChange) => {
+
+          const onFileListChanged = (fileList) => {
+            onChange(fileList);
+            if(model.onFileChanged) executeScript(model.onFileChanged, fileList);
+          };
+
+          const onDownload = (fileList) => {
+            onChange(fileList);
+            if(model.onDownload) executeScript(model.onDownload, fileList);
+          };
+
           return (
             <StoredFilesProvider
               ownerId={Boolean(ownerId) ? ownerId : Boolean(data?.id) ? data?.id : ''}
@@ -94,7 +104,8 @@ const AttachmentsEditor: IToolboxComponent<IAttachmentsEditorProps> = {
               filesCategory={model.filesCategory}
               baseUrl={backendUrl}
               // used for requered field validation
-              onChange={onChange}
+              onChange={onFileListChanged}
+              onDownload={onDownload}
               value={value}
             >
               <CustomFile
@@ -107,7 +118,6 @@ const AttachmentsEditor: IToolboxComponent<IAttachmentsEditorProps> = {
                 allowedFileTypes={model.allowedFileTypes}
                 maxHeight={model.maxHeight}
                 isDragger={model?.isDragger}
-                onFileListChanged={onFileListChanged}
                 downloadZip={model.downloadZip}
                 layout={model.layout}
                 listType={model.listType}
@@ -147,7 +157,8 @@ const AttachmentsEditor: IToolboxComponent<IAttachmentsEditorProps> = {
       ...migrateFormApi.eventsAndProperties(prev),
       onFileChanged: migrateFormApi.withoutFormData(prev?.onFileChanged),
     }))
-    .add<IAttachmentsEditorProps>(6, (prev) => ({ ...prev, listType: !prev.listType ? 'text' : prev.listType })),
+    .add<IAttachmentsEditorProps>(6, (prev) => ({ ...prev, listType: !prev.listType ? 'text' : prev.listType }))
+    .add<IAttachmentsEditorProps>(7, (prev) => ({ ...prev, propertyName: prev.propertyName || '' })),
 };
 
 export default AttachmentsEditor;