+++ shesha-reactjs/src/designer-components/attachmentsEditor/settings.ts
@@ -15,6 +15,15 @@ export const getSettings = () =>
       content: {
         id: 'pnl54bf6-f76d-4139-a850-c99bf06c8b69',
         components: [...new DesignerToolbarSettings()
+          .addContextPropertyAutocomplete({
+            id: nanoid(),
+            propertyName: 'propertyName',
+            label: 'Property name',
+            parentId: 'pnl54bf6-f76d-4139-a850-c99bf06c8b69',
+            validate: {
+              required: true,
+            },
+          })
           .addPropertyAutocomplete({
             id: '5c813b1a-04c5-4658-ac0f-cbcbae6b3bd4',
             propertyName: 'componentName',
@@ -181,6 +190,25 @@ export const getSettings = () =>
               functionName: 'onFileListChanged',
               useAsyncDeclaration: true,
             },
+            availableConstantsExpression: "    return metadataBuilder.object(\"constants\")\r\n        .addAllStandard()\r\n        .addString(\"value\", \"Component current value\")\r\n        .addObject(\"event\", \"Event callback when user input\", undefined)\r\n        .build();"
+
+          })
+          .addCodeEditor({
+            id: '48ff91b3-5fb1-4e1b-a17f-fskj3e2-d32323dxc',
+            propertyName: 'onDownload',
+            label: 'On Download',
+            labelAlign: 'right',
+            parentId: 'root',
+            hidden: false,
+            description: 'Callback that is triggered when a file is downloaded.',
+            validate: {},
+            settingsValidationErrors: [],
+            wrapInTemplate: true,
+            templateSettings: {
+              functionName: 'onDownload',
+              useAsyncDeclaration: true,
+            },
+            availableConstantsExpression: "    return metadataBuilder.object(\"constants\")\r\n        .addAllStandard()\r\n        .addString(\"value\", \"Component current value\")\r\n        .addObject(\"event\", \"Event callback when user input\", undefined)\r\n        .build();"
           })
           .toJson()
         ]