+++ shesha-reactjs/src/providers/storedFiles/actions.ts
@@ -22,6 +22,9 @@ export enum StoredFilesActionEnums {
   DownloadZipError = 'DOWNLOAD_ZIP_ERROR',
   OnFileAdded = 'ON_FILE_ADDED',
   OnFileDeleted = 'ON_FILE_REMOVED',
+  InitializeFileList = 'INITIALIZE_FILE_LIST',
+  UpdateIsDownloadedSuccess = 'UPDATE_IS_DOWNLOADED_SUCCESS',
+  UpdateAllFilesDownloadedSuccess = 'UPDATE_ALL_FILES_DOWNLOADED_SUCCESS',
 
   /* NEW_ACTION_TYPE_GOES_HERE */
 }
@@ -54,16 +57,16 @@ export const uploadFileErrorAction = createAction<IStoredFilesStateContext, ISto
 
 export const deleteFileRequestAction = createAction<IStoredFilesStateContext, string>(
   StoredFilesActionEnums.DeleteFileRequest,
-  (fileIdToDelete) => ({ fileIdToDelete })
+  (fileIdToDelete) => ({ fileId: fileIdToDelete })
 );
 
 export const deleteFileSuccessAction = createAction<IStoredFilesStateContext, string>(
   StoredFilesActionEnums.DeleteFileSuccess,
-  (fileIdToDelete) => ({ fileIdToDelete })
+  (fileIdToDelete) => ({ fileId: fileIdToDelete })
 );
 export const deleteFileErrorAction = createAction<IStoredFilesStateContext, string>(
   StoredFilesActionEnums.DeleteFileError,
-  (fileIdToDelete) => ({ fileIdToDelete })
+  (fileIdToDelete) => ({ fileId: fileIdToDelete })
 );
 
 export const dowloadZipRequestAction = createAction<IStoredFilesStateContext>(
@@ -112,7 +115,22 @@ export const onFileAddedAction = createAction<IStoredFilesStateContext, IStoredF
 
 export const onFileDeletedAction = createAction<IStoredFilesStateContext, string>(
   StoredFilesActionEnums.OnFileDeleted,
-  (fileIdToDelete) => ({ fileIdToDelete })
+  (fileIdToDelete) => ({ fileId: fileIdToDelete })
+);
+
+export const initializeFileListAction = createAction<IStoredFilesStateContext, IStoredFile[]>(
+  StoredFilesActionEnums.InitializeFileList,
+  (fileList) => ({ fileList })
+);
+
+export const updateIsDownloadedByCurrentUser  = createAction<IStoredFilesStateContext, string>(
+  StoredFilesActionEnums.UpdateIsDownloadedSuccess,
+  (fileId) => ({ fileId })
+);
+
+export const updateAllFilesDownloadedByCurrentUser = createAction<IStoredFilesStateContext>(
+  StoredFilesActionEnums.UpdateAllFilesDownloadedSuccess,
+  () => ({})
 );
 
 /* NEW_ACTION_GOES_HERE */